requirements.txt
flask
rsa
waitress
pytaglib
python-dotenv
cryptography
yt_dlp
